requirements.txt
gunicorn
eventlet
Flask-RESTful
requests
Flask-Cors
flask-session
flask
pymongo 
PyJWT
black